requirements.txt
python-telegram-bot[ext]==20.7
asyncpg==0.29.0
aiohttp==3.9.1
APScheduler==3.10.4
pytz==2023.3.post1
python-dotenv==1.0.0